+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added `AsStd140::std140_size` for easily pre-sizing buffers.
 * `Sizer::add` now returns the offset the added value would be written to.
 * Added `std140::DynamicUniform` for aligning dynamic uniform members.
+* Added `Writer::write_slice` for writing multiple values in a row.
 
 ## 0.3.0 (2020-09-22)
 * Added `Std140::as_bytes`, reducing the need to work with bytemuck directly.